Added 409 response.

+++ src/elections/urls.py
@@ -275,15 +275,25 @@ async def update_election(
     "/{election_name}",
     description="Deletes an election from the database. Returns whether the election exists after deletion.",
     response_model=SuccessResponse,
-    responses={401: {"description": "Need to be logged in as an admin.", "model": DetailModel}},
+    responses={
+        401: {"description": "Need to be logged in as an admin.", "model": DetailModel},
+        409: {"description": "Election is still referenced by nominee applications.", "model": DetailModel}
+    },
     operation_id="delete_election",
     dependencies=[Depends(perm_election)],
 )
 async def delete_election(db_session: database.DBSession, election_name: str):
     slugified_name = slugify(election_name)
 
+    existing_applications = await candidates.crud.get_all_candidates_in_election(db_session, slugified_name)
+    if existing_applications: 
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail=f"{election_name} is still referenced by nominee applications.",
+        )
+ 
     await elections.crud.delete_election(db_session, slugified_name)
     await db_session.commit()
-
+    
     old_election = await elections.crud.get_election(db_session, slugified_name)
     return JSONResponse({"success": old_election is None})
